Create comment and like controler in separate file
