+++ app/helpers/comment_helper.rb
@@ -1,0 +1,2 @@
+module CommentHelper
+end
