+++ app/helpers/like_helper.rb
@@ -1,0 +1,2 @@
+module LikeHelper
+end
