+++ spec/requests/comment_request_spec.rb
@@ -1,0 +1,4 @@
+require 'rails_helper'
+
+RSpec.describe 'Comments', type: :request do
+end
