+++ spec/requests/like_request_spec.rb
@@ -1,0 +1,4 @@
+require 'rails_helper'
+
+RSpec.describe 'Likes', type: :request do
+end
